fix: Ensure processed queue is emptied
As the handlers are kept in memory, they keep this list state between processing. This leads to the
occaisional failure when we inevitably end up with more than 10 in the list. We should consider
tests and a cleaner solution, but this will keep the bots uptime and logs quieter for now.