fix: Scheduled task permissions
ECS uses cloudwatch to schedule tasks, and needs to be able to pass through the IAM, to which
the scheduler needs permission to do that. Though the tasks are being renamed back, this prevents
a situation in the future where we add/change a task and wonder why it does not invoke and have
limited ability to diagnose beyond seeing it fail to invoke.