fix: High number of freezes exceeds WCU
We use a limited WCU to keep within the DynamoDB free tier. It in general more than covers our
very minimal needs. However this operation will exceed if there a large number of writes to
commit as the batch has no rate limiting and will occur on context exit if not done during.