fix: LetsEncrypt Cloudflare
As we're no longer using route53, certbot was unable to perform the dns
challenge to renew the certificate. This switches to using dns-cloudflare
to perform a similar function.

The main challenge was that the official container only supports a file or
cli switch for the token. Which is suboptimnal. The awscli container was
added as a dependency to pull the token into a ephemeral container volume
shared with certbot so it can authenticate.

A couple of small work arounds have been added to keep the change set
tightly scoped. But this is prod as it is deployed right now.